+++ src/main/java/com/fakru/interview/tracker/service/InterviewService.java
@@ -1,12 +1,16 @@
 package com.fakru.interview.tracker.service;
 
+import com.fakru.interview.tracker.annotation.LogExecutionTime;
 import com.fakru.interview.tracker.dynamodata.Interview;
+import com.fakru.interview.tracker.exception.VerificationException;
 import com.fakru.interview.tracker.model.request.AddInterviewRequest;
 import com.fakru.interview.tracker.model.response.InterviewResponse;
 import com.fakru.interview.tracker.repository.JobRepository;
+import com.nimbusds.jwt.JWTClaimsSet;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import software.amazon.awssdk.services.dynamodb.model.QueryResponse;
+import util.JoseJwtUtil;
 
 import java.sql.Timestamp;
 import java.util.List;
@@ -15,10 +19,13 @@ import java.util.List;
 public class InterviewService {
 
     private final JobRepository jobRepository;
+    private final VerificationService verificationService;
 
     @Autowired
-    public InterviewService(JobRepository jobRepository) {
+    public InterviewService(JobRepository jobRepository,
+                            VerificationService verificationService) {
         this.jobRepository = jobRepository;
+        this.verificationService = verificationService;
     }
 
     public void addInterview(String userId, String jobId, AddInterviewRequest request) {
@@ -40,5 +47,14 @@ public class InterviewService {
                         .interviewDateTime(new Timestamp(Long.parseLong(item.get("interview_date_time").n()))) // Parse timestamp
                         .build())
                 .toList();
+    }
+
+    @LogExecutionTime
+    public String sendPhoneOTP1(JWTClaimsSet claimsSet, String phoneNumber) {
+        boolean isSent = verificationService.sendPhoneOTP(phoneNumber);
+        if (!isSent) {
+            throw new VerificationException("Error while sending OTP. Please try again after sometime");
+        }
+        return JoseJwtUtil.generateSafeToken(claimsSet.getSubject(), claimsSet.getClaims());
     }
 }
